rename msgIsPaused to msgReceiverQueue for the future build fix for cabal test ghc-specter-daemon
