+++ daemon/test/GHCSpecter/GraphLayout/Algorithm/BFSSpec.hs
@@ -1,12 +1,12 @@
-module GHCSpecter.Util.Graph.BFSSpec (spec) where
+module GHCSpecter.GraphLayout.Algorithm.BFSSpec (spec) where
 
 import Data.Functor.Identity (runIdentity)
 import Data.IntMap qualified as IM
-import GHCSpecter.Util.Graph.BFS
+import GHCSpecter.GraphLayout.Algorithm.BFS
   ( runMultiseedStagedBFS,
     runStagedBFS,
   )
-import GHCSpecter.Util.Graph.Builder (makeBiDep)
+import GHCSpecter.GraphLayout.Algorithm.Builder (makeBiDep)
 import Test.Hspec
   ( Spec,
     describe,
