+++ daemon/test/GHCSpecter/GraphLayout/Algorithm/ClusterSpec.hs
@@ -1,10 +1,10 @@
-module GHCSpecter.Util.Graph.ClusterSpec (spec) where
+module GHCSpecter.GraphLayout.Algorithm.ClusterSpec (spec) where
 
 import Data.IntMap qualified as IM
 import Data.List qualified as L
 import GHCSpecter.Channel.Outbound.Types (ModuleGraphInfo (..))
-import GHCSpecter.Util.Graph.Builder (makeBiDep)
-import GHCSpecter.Util.Graph.Cluster
+import GHCSpecter.GraphLayout.Algorithm.Builder (makeBiDep)
+import GHCSpecter.GraphLayout.Algorithm.Cluster
   ( ClusterState (..),
     ClusterVertex (..),
     degreeInvariant,
